Fix v2

+++ JS/CartController.js
@@ -1,6 +1,118 @@
+var cart = [];
+var allData = [];
+
+let renderProduct = () => {
+    fetch('http://localhost:3000/product')
+    .then(response => {
+        if(!response.ok){
+            throw Error('ERROR');
+        }
+        return response.json();
+    })
+    .then(data => {
+        console.log(data);
+        allData=data;
+        const subHtml = data.slice(0,3).map(item => {
+            return `
+            <div class="product-item" >
+            <div class="product-img">
+                <img src="https://picsum.photos/300/340" alt=""
+                    class="figure-img img-fluid ">
+            </div>
+            <div class="product-description">
+                <p>${item.product_name}</p>
+                <div class="product-price">
+                    $${item.product_price}
+                </div>
+            </div>
+            <div class="item-hover" id="${item.id}">
+                <a href="./productPage.html" class="btn go-to-detail" role="button" >
+                    Detail
+                </a>
+                <button class="btn add-to-cart" >
+                    Add to cart
+                </button>
+            </div>
+            </div>
+            `     
+        }).join("");
+        $("#also-like").append(subHtml);
+
+        const subHtml2 = data.slice(-3).map(item => {
+            return `
+            <div class="product-item" >
+            <div class="product-img">
+                <img src="https://picsum.photos/300/340" alt=""
+                    class="figure-img img-fluid ">
+            </div>
+            <div class="product-description">
+                <p>${item.product_name}</p>
+                <div class="product-price">
+                    $${item.product_price}
+                </div>
+            </div>
+            <div class="item-hover" id="${item.id}">
+                <a href="./productPage.html" class="btn go-to-detail" role="button" >
+                    Detail
+                </a>
+                <button class="btn add-to-cart" >
+                    Add to cart
+                </button>
+            </div>
+            </div>
+            `     
+        }).join("");
+        $("#recently-viewed").append(subHtml2);
+
+    })
+    .then(()=>{
+        $(".add-to-cart").click(function () {
+            let id = $(this).closest('.item-hover').attr('id');
+            
+            var cart = JSON.parse(localStorage.getItem("cart"));
+            if(cart == null) cart = [];
+            let temp = cart.find(i => i.id == id);
+            if(!temp){
+                let item = allData.find(i => i.id == id);
+                var addToCart = {
+                    "id" : item.id,
+                    "product_name" : item.product_name,
+                    "product_price" : item.product_price,
+                    "count" : 1
+                }
+                cart.push(addToCart);
+            }
+            else {
+                cart.map(i => i.id==id?i.count++:i);
+            }
+            console.log(cart);
+            localStorage.setItem("cart",JSON.stringify(cart));
+            alert("Product added !");
+        });
+        $(".go-to-detail").click(function () {
+            let id = $(this).closest('.item-hover').attr('id');
+            sessionStorage.setItem("curent-id",id);
+        })
+    })
+}
+
+let itemChange = () => {
+    let yourCart = 0;
+    let vat = 0;
+    let orderTotal = 0;
+    cart.forEach(item => {
+        yourCart += (item.product_price*item.count);
+        vat = yourCart/10;
+    });
+    orderTotal=yourCart+vat;
+    $("#your-cart").text(yourCart);
+    $("#vat").text(vat);
+    $("#order-total").text(orderTotal);
+}
+
 let renderCart = () => {
     console.log(1);
-    var cart = JSON.parse(localStorage.getItem("cart"));
+    cart = JSON.parse(localStorage.getItem("cart"));
     if(cart == null) cart = [];
     let html = `<div class="row cart-row">
     <div class="col-4 cart-title">
@@ -70,6 +182,7 @@ let renderCart = () => {
         </div>
     </div>`
     }).join("");
+    itemChange();
     $(".cart-container").prepend(html);
     $(".increase-count").click(function (e) { 
         let id = $(this).closest('.cart-row').attr('id');
@@ -87,8 +200,9 @@ let renderCart = () => {
         }
         if(item.count>1) $(this).siblings(".decrease-count").removeClass("disabled");
         cart = cart.map(i => i.id==id?newItem:i);
-        localStorage.setItem("cart",JSON.stringify(cart));
-        console.log(cart);
+        // localStorage.setItem("cart",JSON.stringify(cart));
+        // console.log(cart);
+        itemChange();
     });
     $(".decrease-count").click(function (e) { 
         let id = $(this).closest('.cart-row').attr('id');
@@ -109,8 +223,9 @@ let renderCart = () => {
             }
             if(item.count==1) $(this).addClass("disabled");
         cart = cart.map(i => i.id==id?newItem:i);
-        localStorage.setItem("cart",JSON.stringify(cart));
-        console.log(cart);
+        // localStorage.setItem("cart",JSON.stringify(cart));
+        // console.log(cart);
+        itemChange();
         }
         else if (item.count <= 1) {
             let check = confirm("Do you want to remove this item from cart ?");
@@ -118,7 +233,8 @@ let renderCart = () => {
                 let removeId = "#" + id.toString();
                 cart = cart.filter(i => i.id != id);
                 $(removeId).remove();
-                localStorage.setItem("cart",JSON.stringify(cart));
+                itemChange();
+                // localStorage.setItem("cart",JSON.stringify(cart));
             }
         }
     });
@@ -140,8 +256,9 @@ let renderCart = () => {
         if(newCount>1) $(this).siblings(".decrease-count").removeClass("disabled");
         if(newCount<=1) $(this).siblings(".decrease-count").addClass("disabled");
         cart = cart.map(i => i.id==id?newItem:i);
-        localStorage.setItem("cart",JSON.stringify(cart));
-        console.log(cart);
+        // localStorage.setItem("cart",JSON.stringify(cart));
+        // console.log(cart);
+        itemChange();
     });
     $(".item-delete").click(function (e) { 
         let id = $(this).closest('.cart-row').attr('id');
@@ -150,10 +267,28 @@ let renderCart = () => {
             let removeId = "#" + id.toString();
             cart = cart.filter(i => i.id != id);
             $(removeId).remove();
-            localStorage.setItem("cart",JSON.stringify(cart));
+            itemChange();
+            // localStorage.setItem("cart",JSON.stringify(cart));
         }
+    });
+    $(".update-cart").click(function (e) {
+        let check = confirm("Do you want to update your cart ?");
+        if (check){
+            localStorage.setItem("cart",JSON.stringify(cart));
+            alert("Your cart is now up-to-date");
+        }
+    });
+    $("#checkout").click(function (e) { 
+        let checkout = {
+            "price-total" : $("#your-cart").text(yourCart),
+            "vat" : $("#vat").text(vat),
+            "order-total" : $("#order-total").text(orderTotal),
+            "cart" : cart
+        }
+        localStorage.setItem("checkout",JSON.stringify(checkout));
     });
 }
 
 renderCart();
 
+renderProduct();
